Refactoring to run base and simulation based tests in parallel
