+++ bark_ml/tests/py_library_tf2rl_tests/generate_expert_trajectories_tests/data/BUILD
@@ -1,0 +1,5 @@
+filegroup(
+    name = "generate_expert_trajectories_tests_data",
+    srcs = glob(["expert_trajectories/**", 'interaction_data_set_mock/**']),
+    visibility = ["//visibility:public"],
+)
